reduce number of redundant channels quite a bit
